Added ability to conditionally choose data from either outfit data or state for related products.
