+++ src/redux/actions/getRelatedProducts.js
@@ -1,6 +1,6 @@
 import axios from "axios";
+import {fetchProducts} from "./getProductsHelper.js"
 import { GET_RELATED_PRODUCTS } from "./actionTypes.js";
-// import store from "../store.js"; // DO NOT NORMALLY LOAD STORE <- DONE FOR DEMONSTRATION PURPOSES
 
 export function fetchRelatedProducts(id) {
   return function(dispatch) {
@@ -26,12 +26,4 @@ function getRelatedProducts(data) {
     type: GET_RELATED_PRODUCTS,
     payload: data
   };
-}
-
-export function fetchProducts(ids) {
-  let requests = [];
-  for (let id of ids) {
-    requests.push(axios.get(`http://3.134.102.30/products/${id}`));
-  }
-  return axios.all(requests);
 }
